fix delete

+++ controllers/carController.js
@@ -2,7 +2,11 @@ const CarRepository = require("../repositories/carRepository.js");
 const fs = require("fs");
 const path = require("path");
 const dotenv = require("dotenv");
-const { saveImages, updateImages, deleteImages } = require("../utils/imageUtils.js");
+const {
+  saveImages,
+  updateImages,
+  deleteImages,
+} = require("../utils/imageUtils.js");
 
 dotenv.config();
 
@@ -54,12 +58,24 @@ class CarController {
   }
 
   async delete(req, res) {
-    // Delete the related image files from the file system
-    await deleteImages(car.images);
+    try {
+      // Retrieve the car to get its images
+      const car = await CarRepository.findById(req.params.id);
 
-    // Delete the car from the database
-    await CarRepository.delete(req.params.id);
-    res.json({ message: "Car deleted" });
+      if (!car) {
+        return res.status(404).json({ error: "Car not found" });
+      }
+
+      // Delete the related image files from the file system
+      await deleteImages(car.images);
+
+      // Delete the car from the database
+      await CarRepository.delete(req.params.id);
+      res.json({ message: "Car deleted" });
+    } catch (error) {
+      console.error(error);
+      res.status(500).json({ error: "Internal Server Error" });
+    }
   }
 }
